requirements.txt
pandas
numpy
scikit-learn
xgboost
optuna
scipy
streamlit
pyyaml
